Move Serializable class to types directory and clean up headers
