+++ src/nupic/types/Serializable.hpp
@@ -33,43 +33,37 @@
 #include <capnp/serialize.h>
 #include <kj/std/iostream.h>
 
-using namespace std;
-
 namespace nupic {
-  namespace base {
-    namespace serializable {
 
-      /**
-       * Base Serializable class that any serializable class
-       * should inherit from.
-       */
-      template <class ProtoT>
-      class Serializable {
-        public:
-          void write(ostream& stream) const
-          {
-            capnp::MallocMessageBuilder message;
-            typename ProtoT::Builder proto = message.initRoot<ProtoT>();
-            write(proto);
+  /**
+   * Base Serializable class that any serializable class
+   * should inherit from.
+   */
+  template <class ProtoT>
+  class Serializable {
+    public:
+      void write(std::ostream& stream) const
+      {
+        capnp::MallocMessageBuilder message;
+        typename ProtoT::Builder proto = message.initRoot<ProtoT>();
+        write(proto);
 
-            kj::std::StdOutputStream out(stream);
-            capnp::writeMessage(out, message);
-          }
+        kj::std::StdOutputStream out(stream);
+        capnp::writeMessage(out, message);
+      }
 
-          void read(istream& stream)
-          {
-            kj::std::StdInputStream in(stream);
+      void read(std::istream& stream)
+      {
+        kj::std::StdInputStream in(stream);
 
-            capnp::InputStreamMessageReader message(in);
-            typename ProtoT::Reader proto = message.getRoot<ProtoT>();
-            read(proto);
-          }
+        capnp::InputStreamMessageReader message(in);
+        typename ProtoT::Reader proto = message.getRoot<ProtoT>();
+        read(proto);
+      }
 
-          virtual void write(typename ProtoT::Builder& proto) const = 0;
-          virtual void read(typename ProtoT::Reader& proto) = 0;
-      };
+      virtual void write(typename ProtoT::Builder& proto) const = 0;
+      virtual void read(typename ProtoT::Reader& proto) = 0;
+  };
 
-    } // end namespace serializable
-  } // end namespace base
 } // end namespace nupic
 #endif // NTA_serializable_HPP
